Broke out protected classes from GroupByFunctionBase
